extract label/annotation utilities for reuse
